requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
fuzzywuzzy==0.18.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Levenshtein==0.25.1
MarkupSafe==2.1.5
nltk==3.6.5
numpy==2.0.0
pandas==2.2.2
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.4
regex==2024.5.15
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
Werkzeug==3.0.3

